Move code to release to mallocspace.
